refactor: multi, single slider 한 폴더로 이동, 경로 수정
