+++ src/components/kakao/Slider/SingleRangeSlider/index.tsx
