+++ src/components/kakao/index.ts
@@ -1,4 +1,4 @@
 export { default as KakaoMapContainer } from './Map';
-export { default as MultiRangeSlider } from './MultiRangeSlider';
+export { default as MultiRangeSlider } from './Slider/MultiRangeSlider';
 export { default as FilterModal } from './FilterModal';
-export { default as SingleRangeSlider } from './SingleRangeSlider';
+export { default as SingleRangeSlider } from './Slider/SingleRangeSlider';
